Refactor: Refactoring site-admin
Arranged the file structure properly.
